+++ reservation-service/src/main/java/dunkirk/reservation/dto/ProductForMainDto.java
@@ -1,17 +1,54 @@
-package dunkirk.reservation.dto;
-
-import java.sql.Timestamp;
-
-import dunkirk.reservation.domain.SalesFlag;
+package durkirk.reservation.dto;
 
 public class ProductForMainDto {
 	private int id;
-	private int categoryId;
 	private String name;
 	private String description;
-	private Timestamp salesStart;
-	private Timestamp salesEnd;
-	private SalesFlag salesFlag;
-	private String event;
+	private String placeName;
+	private int fileId;
+	
+	public ProductForMainDto(){
+		
+	}
 
+	public int getId() {
+		return id;
+	}
+
+	public void setId(int id) {
+		this.id = id;
+	}
+
+	public String getName() {
+		return name;
+	}
+
+	public void setName(String name) {
+		this.name = name;
+	}
+
+	public String getDescription() {
+		return description;
+	}
+
+	public void setDescription(String description) {
+		this.description = description;
+	}
+
+	public String getPlaceName() {
+		return placeName;
+	}
+
+	public void setPlaceName(String placeName) {
+		this.placeName = placeName;
+	}
+
+	public int getFileId() {
+		return fileId;
+	}
+
+	public void setFileId(int fileId) {
+		this.fileId = fileId;
+	}
+	
 }
