+++ reservation-service/src/main/webapp/WEB-INF/views/detail.jsp
